// api/model/pouchUpdateDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * POCHY API 명세서
 * POCHY Swagger UI
 * OpenAPI spec version: 1.0.0
 */
import type { CosmeticsDto } from './cosmeticsDto';
import type { WappenDto } from './wappenDto';

export interface PouchUpdateDto {
  theme?: string;
  cosmeticList?: CosmeticsDto[];
  wappenList?: WappenDto[];
}